Datamapper Compatibility Changes
Background:
ActiveRecord and datamapper differ in how they use the bang('!')
charactor in their methods.  AR uses it to for force an exception,
while DM uses it to skip hooks.  Factory_Girl used the save!
method in FG::Proxy::Create#result which cause unexpected loss of
before and after filters in DM application using FG for tests.

Changes:
This commit changes FG::P::Create#result to use the #save method
common to both ORMs.  It then raises it's own error if the save
fails.  Tests now reflect this change as well with generate_spec
looking for the new error message, and create_spec testing for
and stubbing out 'save' instead of 'save!'.